# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Piece-Meal App
# during dev, should point to your local machine
# host: piecemeal-api.herokuapp.com
host: localhost:10010
...
basePath: /api/v1
...
schemes:
  # tip: remove http to make production-grade
  # - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /clients:
    # our controller name
    x-swagger-router-controller: clients
    get:
      description: This endpoint returns all clients available in the database
      # define the type of response for Success "200" and Error
      operationId: getClients
      parameters:
        - name: token
          in: header
          description: Users authentication token
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetUsersResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    post:
      description: Add a new client
      operationId: addClient
      # movie info to be stored
      parameters:
        - name: client
          in: body
          description: client object
          required: true
          schema:
            $ref: "#/definitions/NewClient"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Client"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /clients/{user_id}:
    # our controller name
    x-swagger-router-controller: clients
    get:
      description: Retrieve a single client by id
      operationId: getClient
      # define the type of response for Success "200" and Error
      parameters:
        - name: token
          in: header
          description: Users authentication token
          required: true
          type: string

        - name: user_id
          type: integer
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetUserResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    put:
      description: Update a client by id
      operationId: updateClient
      parameters:
        - name: token
          in: header
          description: Users authentication token
          required: true
          type: string

        - name: user_id
          type: integer
          in: path
          required: true
        - name: client
          in: body
          description: client object
          required: true
          schema:
            $ref: "#/definitions/Client"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    # delete:
    #   description: delete a client
    #   # define the parameters
    #   operationId: deleteClient
    #   parameters:
    #     - name: user_id
    #       description: Client id
    #       type: integer
    #       in: path
    #       required: true
    #   responses:
    #     "200":
    #       description: Success
    #       schema:
    #         $ref: "#/definitions/GeneralResponse"
    #     default:
    #       description: Error
    #       schema:
    #         $ref: "#/definitions/ErrorResponse"
...
  /clients/{user_id}/restrictions:
    x-swagger-router-controller: clients
    get:
      description: get a client's dietary restrictions
      # define the type of response for Success "200" and Error
      operationId: getRestrictions
      parameters:
        - name: token
          in: header
          description: Users authentication token
          required: true
          type: string

        - name: user_id
          type: integer
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetUserRestrictionsResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    post:
      description: add a new ingredient to the client's restrictions
      operationId: addRestriction
      parameters:
        - name: token
          in: header
          description: Users authentication token
          required: true
          type: string

        - name: user_id
          type: integer
          in: path
          description: client's id
          required: true
        - name: ingredient
          description: Request body
          in: body
          required: true
          schema:
            type: object
            required:
              - ingredient_id
            properties:
              ingredient_id:
                description: the ingredient to add
                type: integer
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    patch:
      description: add a new ingredient to the client's restrictions
      operationId: updateRestrictions
      parameters:
        - name: token
          in: header
          description: Users authentication token
          required: true
          type: string

        - name: user_id
          type: integer
          in: path
          description: client's id
          required: true
        - name: ingredients
          in: body
          description: Json object informations.
          required: true
          schema:
            $ref: '#/definitions/IngredientIds'
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    delete:
      description: delete a client's restriction
      operationId: deleteRestriction
      parameters:
        - name: token
          in: header
          description: Users authentication token
          required: true
          type: string

        - name: user_id
          description: Client Id
          type: integer
          in: path
          required: true

        - name: ingredient
          description: Request body
          in: body
          required: true
          schema:
            type: object
            required:
              - ingredient_id
            properties:
              ingredient_id:
                description: the ingredient to delete
                type: integer
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"

        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorResponse"

        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /clients/{user_id}/ingredients/{ingredient_id}/verify:
    x-swagger-router-controller: clients
    get:
      operationId: verifyIngredient
      description: to check if an ingredient is safe for client
      parameters:
        - name: token
          in: header
          description: Users authentication token
          required: true
          type: string

        - name: user_id
          type: integer
          in: path
          required: true

        - name: ingredient_id
          type: integer
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetIngredientVerifyResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /clients/{user_id}/recipes/{recipe_id}/crosscheck:
    x-swagger-router-controller: clients
    get:
      description: to find out what ingredients are safe and non-safe in an existing recipe
      operationId: crossCheckRecipe
      parameters:
        - name: token
          in: header
          description: Users authentication token
          required: true
          type: string

        - name: user_id
          type: integer
          in: path
          required: true

        - name: recipe_id
          type: integer
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetRecipeIngredientSafetyResponse"
        default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
...
  /search/clients:
    x-swagger-router-controller: clients
    get:
      description: get results from client search
      operationId: getUsersSearchResponse
      parameters:
        - name: token
          in: header
          description: Users authentication token
          required: true
          type: string

        - name: first_name
          type: string
          in: query
          required: false
        - name: last_name
          type: string
          in: query
          required: false
        - name: email
          type: string
          in: query
          required: false
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetUsersSearchResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  IngredientIds:
    type: array
    items:
      type: integer
  NewClient:
    type: object
    required:
      - first_name
      - last_name
      - password
      - email
    properties:
      first_name:
        type: string
        description: client object first name
      last_name:
        type: string
        description: client object last name
      email:
        type: string
        description: client object email
      password:
        type: string
        description: client password
  Client:
    type: object
    required:
      - id
      - first_name
      - last_name
      - email
      - is_super_user
    properties:
      id:
        type: integer
        description: client id
      first_name:
        type: string
        description: client object first name
      last_name:
        type: string
        description: client object last name
      email:
        type: string
        description: client object email
      is_super_user:
        type: boolean
        description: does client have super user authorization
...
  Ingredient:
    type: object
    properties:
      id:
        type: number
        description: ingredient id
      name:
        type: string
        description: ingredient object name
      description:
        type: string
        description: description of ingredient
      tags:
        type: array
        items:
          type: string
      alternatives:
        type: array
        items:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
      image_url:
        type: string
        description: link to image of ingredient
    required:
      - name
  GetUserResponse:
    type: object
    properties:
      id:
        type: integer
      first_name:
        type: string
      last_name:
        type: string
      email:
        type: string
      recipes:
        type: array
        items:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            description:
              type: string
            instructions:
              type: array
              items:
                type: object
                properties:
                  step_number:
                    type: integer
                  instructions:
                    type: string
  GetUsersResponse:
    required:
      - clients
    properties:
      # The array of clients
      clients:
        type: array
        items:
          $ref: "#/definitions/Client"
        # type: array
        # items:
        #   type: object
        #   properties:
        #     id:
        #       type: integer
        #     first_name:
        #       type: string
        #     last_name:
        #       type: string
        #     email:
        #       type: string
  GetUsersSearchResponse:
    required:
      - clients
    properties:
      # The array of clients
      clients:
        type: array
        items:
          type: object
          properties:
            id:
              type: integer
            first_name:
              type: string
            last_name:
              type: string

  # GetIngredientSearchResponse:
  #   required:
  #     - ingredients
  #   properties:
  #     # The array of ingredients
  #     ingredients:
  #       type: array
  #       items:
  #         type: object
  #         properties:
  #           id:
  #             type: integer
  #           name:
  #             type: string
  #           tags:
  #             type: array
  #             items:
  #               type: string

  # GetRecipeSearchResponse:
  #   required:
  #     - recipes
  #   properties:
  #     # The array of recipes
  #     recipes:
  #       type: array
  #       items:
  #         type: object
  #         properties:
  #           id:
  #             type: integer
  #           name:
  #             type: string
  #           tags:
  #             type: array
  #             items:
  #               type: string
  GetUserRestrictionsResponse:
    required:
      - ingredients
    properties:
      # The array of restricted ingredients
      ingredients:
        type: array
        items:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            description:
              type: string
            tags:
              type: array
              items:
                type: string
            alternatives:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string

  # PostUserRestrictionResponse:
  #   type: object
  #   properties:
  #     client:
  #       type: object
  #       properties:
  #           id:
  #             type: integer
  #           name:
  #             type: string
  #           restrictions:
  #             type: array
  #             items:
  #               type: object
  #               properties:
  #                 id:
  #                   type: integer
  #                 name:
  #                   type: string
...
  GetIngredientVerifyResponse:
    type: object
    required:
      - safe
    properties:
      safe:
        type: boolean
        description: "Check whether the ingredient is safe or not"
      alternatives:
        type: array
        description: "the alternatives for the ingredient, if it's not safe."
        items:
          $ref: "#/definitions/Ingredient"
  GetRecipeIngredientSafetyResponse:
    # type: string
    type: object
    properties:
      is_safe:
        type: boolean
      forbidden:
        type: array
        items:
          # type: integer
          type: object
          properties:
            id:
              type: integer
            name:
              type: string

  # GetUserRecipeRatingResponse:
  #   required:
  #     - rating
  #   properties:
  #     rating:
  #       type: number
  #       description: "User's rating for a recipe"
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  GeneralResponse:
    type: object
    properties:
      success:
        type: integer
        description: returns 1 if successful
      description:
        type: string
        description: a short comment
    required:
      - success
      - description

  # IngredientResponse:
  #   type: object
  #   properties:
  #     success:
  #       type: integer
  #       description: returns 1 if successful
  #     description:
  #       type: object
  #       properties:
  #       description: a short comment
  #   required:
  #     - success
  #     - description